Rename files and move date formatting to lib
