+++ core/app/c/[communitySlug]/pubs/page.tsx
@@ -2,7 +2,6 @@ import type { Metadata } from "next";
 
 import type { CommunitiesId } from "db/public";
 
-import { PubsDataTable } from "~/app/components/DataTable/PubsDataTable/PubsDataTable";
 import { getPageLoginData } from "~/lib/authentication/loginData";
 import { findCommunityBySlug } from "~/lib/server/community";
 import PubHeader from "./PubHeader";
@@ -34,11 +33,6 @@ export default async function Page(props: Props) {
 
 	return (
 		<>
-			<PubsDataTable
-				communityId={community.id}
-				userId={user.id}
-				searchParams={searchParams}
-			/>
 			<PubHeader communityId={community.id as CommunitiesId} />
 			<PaginatedPubList
 				communityId={community.id}
