+++ core/app/components/DataTable/PubsDataTable/PubsDataTableClient.tsx
